requirements.txt
attrs==19.1.0
backports.functools-lru-cache==1.5
certifi==2018.8.24
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
idna==2.8
ipython-genutils==0.2.0
jsonschema==3.0.1
jupyter-core==4.4.0
kiwisolver==1.0.1
matplotlib==3.0.3
nbformat==4.4.0
numpy==1.14.2
pandas==0.23.4
plotly==3.7.1
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
retrying==1.3.3
scipy==1.2.1
seaborn==0.9.0
six==1.12.0
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
wincertstore==0.2
